# docs/api/instainstru-openapi.yaml
openapi: 3.0.3
info:
  title: InstaInstru API
  description: |
    Backend API for InstaInstru - A platform connecting students with instructors for instant bookings.

    The InstaInstru API provides endpoints for user authentication, instructor profile management,
    availability scheduling, and instant booking of lessons. All endpoints use JSON for requests
    and responses, with JWT Bearer token authentication.
  version: 1.0.0
  contact:
    email: support@instainstru.com
  license:
    name: Proprietary
servers:
  - url: https://api.instainstru.com
    description: Production server
  - url: http://localhost:8000
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/account/suspend:
    post:
      tags:
        - account
      summary: Suspend account
      description: Temporarily suspend the authenticated user's account
      operationId: suspendAccount
      responses:
        '200':
          description: Account suspended successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  suspended_until:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/account/deactivate:
    post:
      tags:
        - account
      summary: Deactivate account
      description: Permanently deactivate the authenticated user's account
      operationId: deactivateAccount
      responses:
        '200':
          description: Account deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/account/reactivate:
    post:
      tags:
        - account
      summary: Reactivate account
      description: Reactivate a suspended or deactivated account
      operationId: reactivateAccount
      responses:
        '200':
          description: Account reactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/account/status:
    get:
      tags:
        - account
      summary: Check account status
      description: Get current account status (active, suspended, deactivated)
      operationId: getAccountStatus
      responses:
        '200':
          description: Account status information
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [active, suspended, deactivated]
                  suspended_until:
                    type: string
                    format: date-time
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Instructors
  /instructors/:
    get:
      tags:
        - instructors
      summary: List all instructors
      description: Get a list of all instructor profiles with their services
      operationId: listInstructors
      security: []
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
          description: Number of records to skip
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            maximum: 100
          description: Maximum number of records to return
      responses:
        '200':
          description: List of instructor profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstructorProfileResponse'
...
  /instructors/{instructor_id}:
    get:
      tags:
        - instructors
      summary: Get instructor by ID
      description: Get a specific instructor's profile by user ID
      operationId: getInstructor
      security: []
      parameters:
        - in: path
          name: instructor_id
          required: true
          schema:
            type: integer
          description: Instructor's user ID
      responses:
        '200':
          description: Instructor profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorProfileResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Availability Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ServiceBase:
      type: object
      required:
        - skill
        - hourly_rate
      properties:
        skill:
          type: string
          minLength: 1
          maxLength: 100
        hourly_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1000
        description:
          type: string
          maxLength: 500
        duration_override:
          type: integer
          minimum: 30
          maximum: 240
    ServiceResponse:
      allOf:
        - $ref: '#/components/schemas/ServiceBase'
        - type: object
          properties:
            id:
              type: integer
            duration:
              type: integer
              description: Effective duration in minutes
    ServiceAreaNeighborhood:
      type: object
      required:
        - neighborhood_id
        - name
        - borough
      properties:
        neighborhood_id:
          type: string
        ntacode:
          type: string
          nullable: true
        name:
          type: string
        borough:
          type: string
    InstructorProfileBase:
      type: object
      required:
        - bio
        - years_experience
      properties:
        bio:
          type: string
          minLength: 10
          maxLength: 1000
        years_experience:
          type: integer
          minimum: 0
          maximum: 50
        min_advance_booking_hours:
          type: integer
          minimum: 0
          maximum: 168
          default: 2
        buffer_time_minutes:
          type: integer
          minimum: 0
          maximum: 60
          default: 0
...
    InstructorProfileResponse:
      allOf:
        - $ref: '#/components/schemas/InstructorProfileBase'
        - type: object
          properties:
            id:
              type: integer
            user_id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
              nullable: true
            user:
              type: object
              properties:
                full_name:
                  type: string
                email:
                  type: string
                  format: email
            service_area_neighborhoods:
              type: array
              items:
                $ref: '#/components/schemas/ServiceAreaNeighborhood'
              default: []
            service_area_boroughs:
              type: array
              items:
                type: string
              default: []
            service_area_summary:
              type: string
              nullable: true
            services:
              type: array
              items:
                $ref: '#/components/schemas/ServiceResponse'

    # Availability Schemas
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
        field:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...